Introduce ReferenceCountedLivenessReferent
