+++ engine/updategraph/src/main/java/io/deephaven/engine/liveness/ReferenceCountedLivenessNode.java
@@ -5,7 +5,6 @@ package io.deephaven.engine.liveness;
 
 import io.deephaven.util.Utils;
 import io.deephaven.util.annotations.VisibleForTesting;
-import io.deephaven.util.referencecounting.ReferenceCounted;
 import org.jetbrains.annotations.NotNull;
 
 import java.lang.ref.WeakReference;
@@ -14,7 +13,7 @@ import java.util.stream.Stream;
 /**
  * {@link LivenessNode} implementation that relies on reference counting to determine its liveness.
  */
-public abstract class ReferenceCountedLivenessNode extends ReferenceCounted implements LivenessNode {
+public abstract class ReferenceCountedLivenessNode extends ReferenceCountedLivenessReferent implements LivenessNode {
 
     final boolean enforceStrongReachability;
 
@@ -43,28 +42,6 @@ public abstract class ReferenceCountedLivenessNode extends ReferenceCounted impl
         if (Liveness.DEBUG_MODE_ENABLED) {
             Liveness.log.info().append("LivenessDebug: Created tracker ").append(Utils.REFERENT_FORMATTER, tracker)
                     .append(" for ").append(Utils.REFERENT_FORMATTER, this).endl();
-        }
-    }
-
-    @Override
-    public final boolean tryRetainReference() {
-        if (Liveness.REFERENCE_TRACKING_DISABLED) {
-            return true;
-        }
-        return tryIncrementReferenceCount();
-    }
-
-    @Override
-    public final void dropReference() {
-        if (Liveness.REFERENCE_TRACKING_DISABLED) {
-            return;
-        }
-        if (Liveness.DEBUG_MODE_ENABLED) {
-            Liveness.log.info().append("LivenessDebug: Releasing ").append(Utils.REFERENT_FORMATTER, this).endl();
-        }
-        if (!tryDecrementReferenceCount()) {
-            throw new LivenessStateException(
-                    getReferentDescription() + " could not be released as it was no longer live");
         }
     }
 
@@ -128,26 +105,9 @@ public abstract class ReferenceCountedLivenessNode extends ReferenceCounted impl
         return true;
     }
 
-    /**
-     * Attempt to release (destructively when necessary) resources held by this object. This may render the object
-     * unusable for subsequent operations. Implementations should be sure to call super.destroy().
-     * <p>
-     * This is intended to only ever be used as a side effect of decreasing the reference count to 0.
-     */
-    protected void destroy() {}
-
     @Override
-    protected final void onReferenceCountAtZero() {
-        if (Liveness.REFERENCE_TRACKING_DISABLED) {
-            throw new IllegalStateException(
-                    "Reference count on " + this + " reached zero while liveness reference tracking is disabled");
-        }
-        try {
-            destroy();
-        } catch (Exception e) {
-            Liveness.log.warn().append("Exception while destroying ").append(Utils.REFERENT_FORMATTER, this)
-                    .append(" after reference count reached zero: ").append(e).endl();
-        }
+    public final void onReferenceCountAtZero() {
+        super.onReferenceCountAtZero();
         tracker.ensureReferencesDropped();
     }
 }
